requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
dash==2.18.1
dash-bootstrap-components==1.4.1
plotly==5.15.0
geopandas==0.13.0
fiona==1.8.21
pyproj>=3.4.1
shapely>=1.8.5
rtree==1.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.6

